% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_missing_zeros.R
\name{add_missing_zeros}
\alias{add_missing_zeros}
\title{Add missing zeros to data frame of catch rates}
\usage{
add_missing_zeros(data_frame, unique_sample_ID_colname, sample_colname,
  species_colname, species_subset = NULL, Method = "Slow",
  if_multiple_records = "Error", verbose = TRUE, save_name = NULL,
  error_tol = 1e-12)
}
\arguments{
\item{data_frame}{data frame containing data that missing zeros}

\item{unique_sample_ID_colname}{column name from 'data_frame' that gives a unique ID for each sample (where rows with the same unique ID are different species from the same sample)}

\item{sample_colname}{column name from 'data_frame' for sample of each species (e.g., representing catch in weight)}

\item{species_colname}{column name from 'data_frame' identifying species sampled for each row}

\item{species_subset}{character vector giving levels of 'species_colname' that should be retained (representing which species to keep);  if NULL, then \code{species_subset=levels(data_frame[,'species_colname'])}}

\item{Method}{whether to use fast or slow method (fast may be more robust, and has more options for multiple_records)}

\item{if_multiple_records}{What to do if multiple rows have the same vales for \code{data_frame[,'unique_sample_ID_colname']} and \code{data_frame[,'species_colname']}}

\item{verbose}{Boolean, whether to print output to screen}

\item{save_name}{optional directory for saving result locally}

\item{error_tol}{value used in checking that final biomass matches original biomass (if \code{error_tol=Inf} then no error checking is done)}
}
\value{
A data frame with number of rows equal to number of unique samples times number of species
}
\description{
\code{add_missing_zeros} takes a compressed `long-form` data frame, and adds rows for samples that were conducted but didn't catch any of a given fish species
}

