% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_growth.R
\docType{data}
\name{adult_growth}
\alias{adult_growth}
\title{Growth data of adult fish}
\format{
A dataset containing 16795 rows and 4 variables
\describe{
\item{fish_code}{Foreign key, matches to `adults`. Identifies the fish being measured. }
\item{period}{a count of each otolith increment. Counts a day in
the fishes life}
\item{position}{the distance of the increment from the centre of the
otolith}
\item{distance}{the distance of the increment from the previous
increment}
}
}
\usage{
adult_growth
}
\description{
A dataset containing daily age and growth data for adult Galaxis maculatus.
}
\examples{
data(adult_growth)

}
\keyword{datasets}
