% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_growths.R
\docType{data}
\name{adult_growths}
\alias{adult_growths}
\title{Growth data of adult fish}
\format{A dataset containing 16795 rows and 6 variables
\describe{
\item{fish_id}{a unique identifier for each fish}
\item{site}{one of two collection sites}
\item{month}{month that the fish was collected on}
\item{increment}{a count of each otolith increment. Counts a day in
the fishes life}
\item{position}{the distance of the increment from the centre of the
otolith}
\item{distance}{the distance of the increment from the previous
increment}
}}
\usage{
adult_growths
}
\description{
A dataset containing adult data for adults of a migratory
fish species. It matches with the "adult_morphologies" dataset.
}
\keyword{datasets}
