% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_metrics.R
\docType{data}
\name{adult_metrics}
\alias{adult_metrics}
\title{Adult fish metrics data}
\format{
A dataset containing 48 rows and 6 variables
\describe{
\item{fish_id}{a unique identifier for each fish}
\item{standard_length}{standard length of the fish (distance from
posterior to caudal peduncle), cm}
\item{body_depth}{body depth of the fish at its maximum point, cm}
\item{age}{Age of fish when caught (days)}
\item{birthdate}{Day fish hatched}
\item{growth_rate}{Average daily growth of fish (mm/day)}
}
}
\usage{
adult_metrics
}
\description{
A dataset containing metrics data for adult Galaxis maculatus.
}
\examples{
data(adult_metrics)
}
\keyword{datasets}
