% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_ERDDAP.R
\name{download_ERDDAP}
\alias{download_ERDDAP}
\title{Download and process satellite data from ERDDAP}
\usage{
download_ERDDAP(Lat_bounds = c(min = 53, max = 66), Lon_bounds = c(min =
  -180, max = -157), Date_bounds = c(min = "06-01", max = "10-01"),
  Year_set = 2010, Variable = "SST", by = c(Lat = 1, Lon = 1, Day = 1))
}
\arguments{
\item{Lat_bounds}{minimum and maximum Latitude boundaries for downloading data}

\item{Lon_bounds}{minimum and maximum Longitude boundaries}

\item{Date_bounds}{minimum and maximum dates (in 'month-day' format)}

\item{Year_set}{Numeric vector of years to query}

\item{Variable}{character giving variable to download (options are 'SST' or 'Chl-a')}
}
\value{
A data frame of ERDDAP data
\describe{
  \item{Lat}{Latitude}
  \item{Lon}{Longitude}
  \item{...}{ERDDAP data columns}
}
}
\description{
\code{Download_ERDDAP} queries the ERDDAP API for global sea-surface temperature and chlorophyll-a with reasonable temperal coverage
}

