% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_growth.R
\docType{data}
\name{juvenile_growth}
\alias{juvenile_growth}
\title{Growth data of juvenile fish}
\format{
A dataset containing 87581 rows and 5 variables
\describe{
\item{fish_code}{Foreign key, links with `juveniles`}
\item{otolith_code}{Alternate key}
\item{period}{a count of each otolith increment. Counts a day in
the fishes life}
\item{position}{the distance of the increment from the centre of the
otolith}
\item{distance}{the distance of the increment from the previous
increment}
}
}
\usage{
juvenile_growth
}
\description{
A dataset containing daily growth data for juvenile Galaxis maculatus.
}
\examples{
data(juvenile_growth)
}
\keyword{datasets}
