% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juvenile_morphologies.R
\docType{data}
\name{juvenile_morphologies}
\alias{juvenile_morphologies}
\title{Morphological data of juvenile fish}
\format{A dataset containing 496 rows and 7 variables
\describe{
\item{fish_id}{a unique identifier for each fish}
\item{site}{one of two collection sites}
\item{day}{day that the fish was collected on}
\item{month}{month that the fish was collected on}
\item{standard_length}{standard length of the fish (distance from
posterior to caudal peduncle)}
\item{body_depth}{body depth of the fish at its maximum point}
\item{otolith_id}{unique identifier for the fishes otolith}
}}
\usage{
data(juvenile_morphologies)
}
\description{
A dataset containing morphological data for juveniles of a migratory
fish species. It matches with the "juvenile_growths" dataset
}
\keyword{datasets}
