% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sol274}
\alias{sol274}
\title{North Sea sole stock assessment input and output (FLStock)}
\format{
FLStock
}
\usage{
data(sol274)
}
\description{
An object of class \emph{FLStock} from the ICES WGNSSK 2021 stock assessment of
sole (\emph{Solea solea}). Conatins estimates of abundance and fishing
mortality-at-age from the model run conducted using AAP (Aarts and Poos).
}
\references{
\url{https://doi.org/10.17895/ices.pub.8211}
}
\keyword{datasets}
